requirements.txt
XlsxWriter==3.2.3
diskcache==5.6.3
pandas==2.2.3
python-pptx==1.0.2
snowflake-connector-python==3.15.0
streamlit